Implement FlexCel in TMSExcelAPI
This commit introduces the use of the FlexCel library to the TMSExcelAPI. It replaces the basic 'Hello World' console output with Excel spreadsheet generation using FlexCel. Now, the program will create an Excel file, add some values, formulas and then save it. The FlexCel package has also been added to the project dependencies.